+++ db/models/processing_status.go
@@ -1,0 +1,9 @@
+package models
+
+const (
+	ProcessingStatusUploaded   = "Uploaded"
+	ProcessingStatusProcessing = "Processing"
+	ProcessingStatusProcessed  = "Processed"
+	ProcessingStatusMatched    = "Matched"
+	ProcessingStatusFailed     = "Failed"
+)