+++ services/services.go
@@ -14,11 +14,11 @@ type Services struct {
 
 func NewServices(repos *repositories.Repos, dependencies *dependencies.DependencyService) *Services {
 	file := NewFileService(dependencies.GetFileStore())
-	imageProcessingService := NewImageProcessingService(repos.TyreImpression, file, repos.File, dependencies.GetFileStore())
+	imageProcessingService := NewImageProcessingService(repos.TyreImpression, repos.TyreModel, file, repos.File, dependencies.GetFileStore())
 
 	return &Services{
 		TyreImpression:  NewTyreImpressionService(repos.TyreImpression, repos.File, dependencies.GetFileStore(), NewUploadedFileValidator(), imageProcessingService),
-		TyreModel:       NewTyreModelService(repos.TyreModel, repos.File, dependencies.GetFileStore(), NewUploadedFileValidator()),
+		TyreModel:       NewTyreModelService(repos.TyreModel, repos.File, dependencies.GetFileStore(), NewUploadedFileValidator(), imageProcessingService),
 		ImageProcessing: imageProcessingService,
 		File:            file,
 	}